test(lexical-graph): share the opensearchpy mock bootstrap across test files
test_opensearch_clients.py, test_opensearch_nextgen.py, and
test_opensearch_vector_indexes.py each defined their own private
NotFoundError/RequestError fake classes and injected them into sys.modules
before importing opensearch_vector_indexes. Since that module binds
`from opensearchpy.exceptions import ...` once at first import, whichever
file's mocks installed first silently won for the rest of the test process --
`except RequestError` in the source would stop matching instances raised by
a sibling file's now-losing class, an order-dependent failure confirmed
reproducible with `pytest test_opensearch_nextgen.py test_opensearch_clients.py`.

Extracts the bootstrap into _opensearch_test_support.py so all three files
share the same FakeNotFoundError/FakeRequestError classes regardless of
collection order.
